rename services and fixtures and spec files
