+++ fixtures/chain-fixtures.fixture.ts
@@ -4,4 +4,4 @@ import { sauceTest } from './saucedemo-test-data.fixture';
 
 export const test = mergeTests(pageObjectTest, sauceTest);
 export { expect } from '@playwright/test';
-export { HamburgerMenu } from '../services/components/HamburgerMenu';
+export { HamburgerMenu } from '../services/components/hamburger-menu';
